chore: refactor some comps
